requirements.txt
pip

aiohttp==3.5.4
mysql-connector-python==8.0.15
envparse==0.2.0